requirements_api.txt
flask
pillow
protobuf
python-dotenv
sentencepiece
tiktoken
torch
transformers
